+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Add `distro_of_tag` and `generate_dockerfile` to `Dockerfile_distro`.
 * Add `nano` to images to satisfy `opam pin` going interactive.
 * Also include `4.03.0` flambda build.
+* Add ARMv7hf Raspbian distro (Wheezy and Jessie).
 
 1.2.0:
 * Add `dev-repo` metadata to OPAM file.